requirements.txt
attrs==21.4.0
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==2.2.1
eventlet==0.33.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
prettytable==3.2.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytest-repeat==0.9.1
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
wcwidth==0.2.5
